pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chat-over-pdf"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.3.21",
    "llama-index-core>=0.12.40",
    "faiss-cpu",
    "pypdf",
    "streamlit>=1.45",
    "python-dotenv",
    "pytesseract",
    "pdf2image",
]

[tool.setuptools.packages.find]
where = ["src"]
